Работа с модулями, создан отдельный файл который генерирует массив
